+++ src/Haskellorls/Formatter/Layout/Grid.hs
@@ -4,7 +4,7 @@ module Haskellorls.Formatter.Layout.Grid
   )
 where
 
-import qualified Data.List as L
+import Data.List (intersperse, transpose)
 import qualified Data.Text as T
 import qualified Data.Text.Lazy as TL
 import qualified Data.Text.Lazy.Builder as TL
@@ -18,7 +18,7 @@ import qualified Haskellorls.WrappedText as WT
 -- > horizontalSplitInto 2 [1 .. 6] == [[1, 3, 5], [2, 4, 6]]
 -- > horizontalSplitInto 3 [1 .. 10] == [[1, 4, 7, 10], [2, 5, 8], [3, 6, 9]]
 verticalSplitInto :: Int -> [a] -> [[a]]
-verticalSplitInto n = L.transpose . slice n
+verticalSplitInto n = transpose . slice n
 
 horizontalSplitInto :: Int -> [a] -> [[a]]
 horizontalSplitInto = splitInto
@@ -51,31 +51,30 @@ slice n xs
   where
     (h, t) = splitAt n xs
 
-splitIntoMaxSum :: Int -> [Int] -> [[Int]]
-splitIntoMaxSum _ [] = []
-splitIntoMaxSum n xs = ys : splitIntoMaxSum n xs'
-  where
-    ys = takeWhileWithAccum (\a b -> a + b <= n) (+) 0 xs
-    xs' = drop (length ys) xs
-
--- | 'takeWhileWithAccum' @predicate accumulator a0 xs@ is similar 'takeWhile'.
--- But this has accumulator and the accumulated value can be used in predicate
--- statement.
+-- | 'takeWhileAccumL' @predicate accumulator acc xs@ is similar 'takeWhile'.
+-- In addition, this function has an accumulator and an accumulated value,
+-- which can be used in predicate statement.
 --
--- > takeWhileWithAccum (\a b -> a + b <= n) (+) 10 [1 .. 10] == [1, 2, 3, 4]
-takeWhileWithAccum :: (a -> b -> Bool) -> (a -> b -> a) -> a -> [b] -> [b]
-takeWhileWithAccum _ _ _ [] = []
-takeWhileWithAccum p accumulator a0 (x : xs) =
-  if p a0 x
-    then x : takeWhileWithAccum p accumulator (accumulator a0 x) xs
+-- >>> takeWhileAccumL (\a b -> a + b <= n) (+) 10 [1 .. 10]
+-- [1, 2, 3, 4]
+takeWhileAccumL :: (a -> b -> Bool) -> (a -> b -> a) -> a -> [b] -> [b]
+takeWhileAccumL _ _ _ [] = []
+takeWhileAccumL p accumulator acc (x : xs) =
+  if p acc x
+    then x : takeWhileAccumL p accumulator (accumulator acc x) xs
     else []
 
 buildValidCommaSeparatedGrid :: Int -> [[WT.WrappedText]] -> [[[WT.WrappedText]]]
-buildValidCommaSeparatedGrid n sss = splitsAt validLengths sss'
+buildValidCommaSeparatedGrid n sss = intersperse [WT.deserialize " "] <$> splitsAt validLengths sss'
   where
-    sss' = mapToInit (<> [WT.deserialize commaSuffix]) sss
-    lengths = map (sum . map termLength) sss'
-    validLengths = map length $ splitIntoMaxSum n lengths
+    sss' = mapToInit (<> [WT.deserialize ","]) sss
+    lengths = sum . map termLength <$> sss'
+    validLengths = length <$> f lengths
+
+    f [] = []
+    f (x : xs) =
+      let l = length $ takeWhileAccumL (\acc y -> acc + y + 1 <= n) (\acc y -> acc + y + 1) x xs
+       in (x : take l xs) : f (drop l xs)
 
 -- | 'splitsAt' @ns xs@ returns splited list which has lists as elements. The
 -- element's length may be not same.
@@ -87,7 +86,7 @@ splitsAt [] _ = []
 splitsAt _ [] = []
 splitsAt (n : nx) xs = ys : splitsAt nx xs'
   where
-    (ys, xs') = L.splitAt n xs
+    (ys, xs') = splitAt n xs
 
 buildValidGrid :: Config.Config -> Int -> [[WT.WrappedText]] -> [[[WT.WrappedText]]]
 buildValidGrid _ _ [] = []
@@ -133,7 +132,7 @@ validateGrid n grid
     maxLen = maximum . map (termLength . TL.toStrict . TL.toLazyText) $ renderGridAsPlain grid
 
 buildGridWithTab :: Config.Config -> Int -> [[WT.WrappedText]] -> [[[WT.WrappedText]]]
-buildGridWithTab config n wtss = map (buildRow (Config.tabSize config) maxLengths) $ L.transpose grid
+buildGridWithTab config n wtss = map (buildRow (Config.tabSize config) maxLengths) $ transpose grid
   where
     grid = splitter n wtss
     maxLengths = mapToInit (+ 2) $ calcEachRowMaxWTLength grid
@@ -191,7 +190,7 @@ buildPaddingsWithSpace len (n : ns) (m : ms) = T.replicate (nLen - mLen) " " : b
     mLen = len + m
 
 buildGridWithSpace :: Config.Config -> Int -> [[WT.WrappedText]] -> [[[WT.WrappedText]]]
-buildGridWithSpace config n wtss = map (buildRowWithSpace maxLengths) $ L.transpose grid
+buildGridWithSpace config n wtss = map (buildRowWithSpace maxLengths) $ transpose grid
   where
     grid = splitter n wtss
     maxLengths = mapToInit (+ 2) $ calcEachRowMaxWTLength grid
@@ -208,6 +207,3 @@ mapToInit f = foldr (\x acc -> if null acc then [x] else f x : acc) mempty
 
 calcEachRowMaxWTLength :: [[[WT.WrappedText]]] -> [Int]
 calcEachRowMaxWTLength = map $ maximum . (0 :) . map (sum . map termLength)
-
-commaSuffix :: T.Text
-commaSuffix = ", "